requirements.txt
fastapi
uvicorn
tensorflow
pandas
google-generativeai
python-dotenv
pydantic
scikit-learn
numpy
